// src/models/_db.ts
export interface _IArticle {
    id: number;
    courseId: number;
    title: string;
    content: string;
    publishAt?: number;
    createAt: number;
    status: number;
    audioId: number;
    contentType: string;

}

export interface _IArticleComment {
    id: number;
    articleId: number;
    accountId: number;
    content: string;
    createAt: number;
    status: number;
    replyTo?: number;

}